split scheduler priorities
